tradester_backend/pyproject.toml
[tool.poetry]
name = "tradester_backend"
version = "0.1.0"
description = ""
authors = ["Jaffar Alzeidi <jaffar.sa139@gmail.com>"]
readme = "README.md"
packages = [{include = "tradester_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
gunicorn = "^20.1.0"
docutils = "^0.19"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^1.0.0"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"
djangorestframework-simplejwt = "^5.2.2"
python-status = "^1.0.1"
django-crontab = "^0.7.1"
apscheduler = "^3.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
